IEM: Windows build fix
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@41837 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMAll/IEMAll.cpp
@@ -7940,7 +7940,7 @@ VMMDECL(VBOXSTRICTRC)       IEMExecOneWithPrefetchedByPC(PVMCPU pVCpu, PCPUMCTXC
         && pCtx->rip == OpcodeBytesPC)
     {
         iemInitDecoder(pIemCpu);
-        pIemCpu->cbOpcode = RT_MIN(cbOpcodeBytes, sizeof(pIemCpu->abOpcode));
+        pIemCpu->cbOpcode = (uint8_t)RT_MIN(cbOpcodeBytes, sizeof(pIemCpu->abOpcode));
         memcpy(pIemCpu->abOpcode, pvOpcodeBytes, pIemCpu->cbOpcode);
         rcStrict = VINF_SUCCESS;
     }
